+++ src/core/apiCodes.js
@@ -7,3 +7,5 @@ export const searchAPI = "https://api.themoviedb.org/3/search";
 export const movieID = "https://api.themoviedb.org/3/movie/";
 export const apiGenres = "https://api.themoviedb.org/3/genre/movie/list";
 export const apiMovieDetails = "movie/";
+export const imgBaseBackdrop = "https://image.tmdb.org/t/p/original";
+export const details = "https://api.themoviedb.org/3/movie/{movie_id}";
