+++ src/features/movies/MovieDetails/getMovieDetails.js
@@ -1,7 +1,10 @@
-import { key, url,} from "../../../core/apiCodes";
+import axios from "axios";
+ const url = "https://api.themoviedb.org/3";
+ const key = "ae7d4af255a05506e1ed3b49e48b0d5c";
 
-export const getMovieDetails = async ({ movieId }) => {
-  const response = await fetch(
+
+export const getMovieDetails = async ( movieId ) => {
+  const response = await axios.get(
     `${url}/movie/${movieId}?api_key=${key}&language=en-US`
   );
   if (!response.ok) {
@@ -11,9 +14,8 @@ export const getMovieDetails = async ({ movieId }) => {
   return await response.data;
 };
 
-
-export const getMovieCredits = async ({ movieId }) => {
-  const response = await fetch(
+export const getMovieCredits = async (movieId ) => {
+  const response = await axios.get(
     `${url}/movie/${movieId}/credits?api_key=${key}&language=en-US`
   );
   if (!response.ok) {
